Refactored Apollo Client, Loading text added
