+++ src/apollo/client/queries/index.ts
@@ -1,0 +1,6 @@
+export * from "./user";
+export * from "./post";
+export * from "./comment";
+export * from "./image";
+export * from "./follow";
+export * from "./like";
